test_Web_Aplication/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beanie==1.23.3
motor==3.7.1
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.25.0
<<<<<<< HEAD
email-validator>=2.2.0
=======
>>>>>>> 02675bc (After Deploy Shamim)

